requirements.txt
pyyaml
telepot